requirements.txt
cachetools==6.0.0
chardet==5.2.0
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pyproject-api==1.9.1
pytesseract==0.3.13
virtualenv==20.31.2
